Move things outside of pyarrow.py
